+++ src/components/AnimateSharedLayout/types.ts
@@ -1,3 +1,4 @@
+import { VisualElement } from "../../render/types"
 import { Transition } from "../../types"
 import { AxisBox2D } from "../../types/geometry"
 import { MotionValue } from "../../value"
@@ -42,4 +43,32 @@ export interface SharedLayoutAnimationConfig {
     crossfadeOpacity?: MotionValue<number>
     shouldStackAnimate?: boolean
     onComplete?: () => void
+}
+
+/**
+ * Handlers for batching sync layout lifecycles. We batches these processes to cut
+ * down on layout thrashing
+ */
+export interface SyncLayoutLifecycles {
+    measureLayout: (child: VisualElement) => void
+    layoutReady: (child: VisualElement) => void
+    parent?: VisualElement
+}
+
+/**
+ * The exposed API for children to add themselves to the batcher and to flush it.
+ */
+export interface SyncLayoutBatcher {
+    add: (child: VisualElement) => void
+    flush: (handler?: SyncLayoutLifecycles) => void
+}
+
+/**
+ * Extra API methods available to children if they're a descendant of AnimateSharedLayout
+ */
+export interface SharedLayoutSyncMethods extends SyncLayoutBatcher {
+    syncUpdate: (force?: boolean) => void
+    forceUpdate: () => void
+    register: (child: VisualElement) => void
+    remove: (child: VisualElement) => void
 }
