+++ src/motion/utils/use-visual-element.ts
@@ -1,7 +1,7 @@
 import { useContext, useEffect } from "react"
-import { PresenceContext } from "../../components/AnimatePresence/PresenceContext"
+import { PresenceContext } from "../../context/PresenceContext"
 import { isPresent } from "../../components/AnimatePresence/use-presence"
-import { LayoutGroupContext } from "../../components/AnimateSharedLayout/LayoutGroupContext"
+import { LayoutGroupContext } from "../../context/LayoutGroupContext"
 import { MotionProps } from "../../motion"
 import { useVisualElementContext } from "../../context/MotionContext"
 import { useSnapshotOnUnmount } from "../../motion/features/layout/use-snapshot-on-unmount"
